refactor: rename submodules
